refactor: split massive index file into logical smaller files. no logic lost, just making this easier to work in. added a platform folder for future platform seperation
