+++ back/src/api.ts
@@ -1,0 +1,267 @@
+import url from "url";
+import axios, { AxiosResponse } from "axios";
+import WebSocket from "ws";
+import { WSS_PORT, WS_FORCE_SYNC_TIME, WS_HB_TIME, WS_MSG_BURST, WS_MSG_RATE, CLIENT_ID, ALLOWED_USERS } from "./config";
+import { TimerWebSocket } from "./types";
+import { bus, emitSync } from "./bus";
+import { usersModel, dbCreate, USER_TABLE } from "./db";
+import { DEFAULT_RATES, normalizeRates } from "./rates";
+import { getUserSession, loginUser, logoutUser, connectTwitchFor, connectStreamlabsFor } from "./session";
+import { setEndTime } from "./timer";
+import { logTimerEvent, sendLogPage } from "./log";
+
+let wss: WebSocket.Server;
+
+function wsCloseError(ws: TimerWebSocket, reason: string){
+    ws.send(
+        JSON.stringify({
+            success: false,
+            error: reason,
+        })
+    );
+    ws.close();
+}
+
+function wsSync(ws: TimerWebSocket) {
+    if (!ws.isReady)
+        return;
+    if (ws.readyState !== WebSocket.OPEN)
+        return;
+    const curSession = getUserSession(ws.userId);
+    ws.send(
+        JSON.stringify({
+            success: true,
+            endTime: curSession.endTime,
+            slStatus: curSession.slStatus,
+            twitchStatus: curSession.twitchStatus,
+            cap: curSession.shouldCap,
+            anon: curSession.ignoreAnon,
+            rates: curSession.rates,
+            connections: {
+                twitch: { channel: curSession.connections.twitch.channel },
+                streamlabs: { hasToken: !!curSession.connections.streamlabs.token }
+            },
+            merchValues: curSession.merchValues
+        })
+    );
+}
+
+async function wsLogin(ws: TimerWebSocket, accessToken: string){
+    ws.userId = 0;
+    ws.isAlive = true;
+    ws.msgTokens = WS_MSG_BURST;
+    ws.msgLast = Date.now();
+    ws.msgWarnAt = 0;
+    ws.forceSyncInterval = setInterval(()=>wsSync(ws),WS_FORCE_SYNC_TIME);
+    ws.hbInterval = setInterval(()=>{
+        if (ws.isAlive == false){
+            wsCloseError(ws, "Did not heartbeat in time!");
+            return;
+        }
+        ws.isAlive = false;
+        ws.ping();
+    }, WS_HB_TIME);
+
+    if (!accessToken){
+        wsCloseError(ws, "No token provided!");
+        return;
+    }
+
+    let userName = "";
+
+    if (CLIENT_ID == ""){ // Unauthorized Logins
+        ws.userId = 1;
+        userName = accessToken;
+    }
+    else { // Authorized Logins
+        let httpRes: AxiosResponse;
+        try {
+            httpRes = await axios.get(`https://api.twitch.tv/helix/users`, {
+                headers: {
+                    Authorization: `Bearer ${accessToken}`,
+                    "Client-Id": CLIENT_ID,
+                }
+            });
+        } catch {
+            wsCloseError(ws, "Failed to login!");
+            return;
+        }
+        ws.userId = httpRes.data.data[0].id;
+        userName = httpRes.data.data[0].login;
+    }
+
+    if (!ALLOWED_USERS.includes(userName)){
+        wsCloseError(ws,"Channel name is not in ALLOWED_USERS!");
+        return;
+    }
+
+    const res = await usersModel.findByPk(ws.userId);
+
+    if (!res){
+        const newUser = {
+            userId: ws.userId,
+            name: userName as string,
+            accessToken: accessToken as string,
+            slToken: "",
+            subTime: USER_TABLE.subTime.defaultValue,
+            dollarTime: USER_TABLE.dollarTime.defaultValue,
+            endTime: USER_TABLE.endTime.defaultValue,
+            shouldCap: USER_TABLE.shouldCap.defaultValue,
+            ignoreAnon: USER_TABLE.ignoreAnon.defaultValue,
+            rates: DEFAULT_RATES,
+            connections: { twitch: { channel: userName }, streamlabs: { token: "" } }
+        }
+        await dbCreate(newUser);
+        loginUser(newUser);
+        return;
+    }
+
+    const curSession = getUserSession(ws.userId);
+
+    if (curSession.userId == 0){
+        wsCloseError(ws, `User ID ${ws.userId} is not logged in but is present in the database!`);
+        return;
+    }
+
+    if (ws.userId == 1 && accessToken != curSession.name){ // Changing unauthorized user's name
+        console.log(`Changing userId 1 from ${curSession.name} to ${accessToken}`);
+        let newSession = Object.assign({},curSession);
+        newSession.name = accessToken;
+        newSession.accessToken = accessToken;
+        logoutUser(1);
+        loginUser(newSession);
+    }
+
+    return;
+}
+
+export function startApi(){
+    wss = new WebSocket.Server({port: WSS_PORT});
+
+    bus.on("sync", (id: number) => {
+        const clientsArr = Array.from(wss.clients);
+        for (let i = 0; i < clientsArr.length; i++){
+            const ws = clientsArr[i] as TimerWebSocket;
+            if (id != ws.userId)
+                continue;
+            // one dying socket must not abort the broadcast to the rest, nor unwind back into the event handler
+            try {
+                wsSync(ws);
+            } catch (err) {
+                console.log("Failed to sync a client:", err);
+            }
+        }
+    });
+
+    bus.on("logEntry", (id: number, entry: any) => {
+        const clientsArr = Array.from(wss.clients);
+        for (let i = 0; i < clientsArr.length; i++){
+            const ws = clientsArr[i] as TimerWebSocket;
+            if (id != ws.userId || ws.readyState !== WebSocket.OPEN)
+                continue;
+            try {
+                ws.send(JSON.stringify({ logEntry: entry }));
+            } catch (err) {
+                console.log("Failed to send a log entry to a client:", err);
+            }
+        }
+    });
+
+    wss.on("connection", (ws: TimerWebSocket, req: any) => {
+        console.log("A client has connected to the WSS backend!");
+        ws.isReady = false;
+
+        const urlParams = url.parse(req.url, true).query;
+        const accessToken = urlParams.token as string;
+
+        wsLogin(ws, accessToken).then(()=>{
+            ws.isReady = true;
+            emitSync(ws.userId);
+        }).catch((err)=>{
+            console.log("wsLogin failed:", err);
+            wsCloseError(ws, "Failed to login!");
+        });
+
+        ws.on("pong",()=>{
+            ws.isAlive = true;
+        });
+
+        ws.on('close',()=>{
+            console.log("A client has disconnected from the WSS backend!");
+            clearInterval(ws.forceSyncInterval);
+            clearInterval(ws.hbInterval);
+        });
+
+        ws.on("message", (data: any)=>{
+            if (!ws.isReady)
+                return;
+            // token-bucket rate limit per connection — guards against FE loops spamming the server
+            const now = Date.now();
+            ws.msgTokens = Math.min(WS_MSG_BURST, ws.msgTokens + ((now - ws.msgLast) / 1000) * WS_MSG_RATE);
+            ws.msgLast = now;
+            if (ws.msgTokens < 1){
+                if (now - ws.msgWarnAt > 5000){
+                    ws.msgWarnAt = now;
+                    console.log(`Rate limiting userId ${ws.userId} (too many messages) — dropping.`);
+                }
+                return;
+            }
+            ws.msgTokens -= 1;
+            const id = ws.userId;
+            const curSession = getUserSession(id);
+
+            try {
+                var jData = JSON.parse(data);
+            } catch (error) {
+                wsCloseError(ws, "Error while parsing JSON payload!");
+                return;
+            }
+
+            switch (jData.event) {
+                case "getTime":
+                    break;
+                case "getLogPage":
+                    sendLogPage(ws, jData.before);
+                    return;
+                case "setConnection": {
+                    const platform = jData.platform;
+                    const config = jData.config || {};
+                    if (platform === "twitch") {
+                        const channel = (typeof config.channel === "string" ? config.channel : "").trim().toLowerCase();
+                        curSession.connections.twitch.channel = channel;
+                        if (curSession.conTMI)
+                            curSession.conTMI.disconnect();
+                        curSession.twitchStatus = false;
+                        curSession.conTMI = channel ? connectTwitchFor(curSession) : undefined;
+                    } else if (platform === "streamlabs") {
+                        if (typeof config.token !== "string" || config.token.length >= 1000)
+                            break;
+                        curSession.connections.streamlabs.token = config.token;
+                        if (curSession.conSL)
+                            curSession.conSL.disconnect();
+                        curSession.slStatus = false;
+                        curSession.conSL = config.token ? connectStreamlabsFor(curSession) : undefined;
+                    }
+                    break;
+                }
+                case "setRates":
+                    curSession.rates = normalizeRates(jData.rates);
+                    break;
+                case "setEndTime": {
+                    const oldET = curSession.endTime;
+                    setEndTime(curSession, Math.trunc(parseInt(jData.value) || 0));
+                    logTimerEvent(curSession, "Manual change", oldET, curSession.endTime);
+                    break;
+                }
+                case "setCap":
+                    curSession.shouldCap = Boolean(jData.value) || false;
+                    setEndTime(curSession, curSession.endTime);
+                    break;
+                case "setAnon":
+                    curSession.ignoreAnon = Boolean(jData.value) || false;
+                    break;
+            }
+            emitSync(id);
+        });
+    });
+}
