+++ back/tsconfig.json
@@ -17,5 +17,5 @@
 		"jsx": "react-jsx",
 		"outDir": "./build"
 	},
-	"include": ["src/*"]
+	"include": ["src/**/*"]
 }
